requirements.txt
clams-python==1.3.1
ffmpeg-python
nemo_toolkit[asr]>=1.23.0
